Move licensing documentation to a standalone file
To save on paper and space, I've moved the GNU FDL to a separate pdf and
taken it out of the individual documents.  This seems consistent with
Section 6 of the license (Collections of Documents), since this entire
repository is best viewed as a single collection.  I've commented out the
LaTeX code to include the license in the individual pdfs, so if anyone
wants to distribute any of them on their own it should be very easy to
reintroduce the text and comply with the FDL.