Add automatic generation of citation info
This commit adds a command to the Makefile to generate a BibTeX record for the
book; it also adds that information to the front page of each document so that
people will (hypothetically) know how to cite it.  I expect the lack of author
and the versioning info to confuse most people.