Add placeholder VERSION.tex and CITATION.bib files
This commit adds templates for VERSION.tex and CITATION.bib: those
files are automatically generated by the makefile to contain the
date and version of the current commit.  Unfortunately, this means
that they do not exist when the repository is first cloned, which
causes errors in LaTeX.  So this commit adds files that explain
that they are just placeholders and gives directions for how to
determine the correct information.  I've also added those directions
to the README file.