Move FDL and author list into the individual pdf files
In earlier versions, we viewed the collection of pdfs as a single
entity and encouraged the FDL and author lists to be distributed
as separate pdf files (to minimize space and duplication).  But
it does seem likely that people will encounter these pdfs separately,
so we've decided to include the license and author list in every
pdf by default.

I still think that our earlier view is correct and that the license
and author list can be distributed once (for example, if the book
is used for a class, the instructor can compile the pdfs without
licensing information, but just give the license to the students
the first day), so this is just a matter of emphasis.