Move tufte-latex files to `external` folder
This commit sets up a folder labeled `external` that will contain all of
the files external to the project.  Separating them will make it easier to
attribute the correct license, especially if we need to add local copies
of more packages or fonts.

The commit also renames the `tex` folder to the more descriptive name
`common`.