Move project information into separate section
This commit moves information about the EFLP project off
the front page and into a separate section of the texts.
The project information has gotten a little long, so it
was pushing the table of contents off the front.  The
notice that the document is licensed under the FDL and the
copyright notice are both still on the front page.