Edit and move GLS material
This commit moves the GLS material from the linear regression
document; "infeasible" GLS is moved to the section on finite
sample properties of the OLS estimator and "feasible" GLS is
now in the asymptotics document with the rest of the discussion
of heteroskedasticity.

Both sections have been tightened up and put into paragraphs
and sentences.