Add more error-tolerant compilation options to makefile
This commit makes the Makefile a little smarter; it checks for
the existence of latexmk and, if it's not there, calls latex
and bibtex repeatedly.

The commit also explicitly deletes files in `burn` and `clean`.

I first saw the latexmk trick in the makefile for *Homotopy Type
Theory: Univalent Foundations of Mathematics* and this is a minor
variation of their approach.