Rewrite latex inclusion of generated files
This commit adds new commands to give default values of the
citation information, version number, and figures.  These
files are generated by make and may not be present if someone
opens core_econometrics.tex and starts hacking; now that will
work, but there will be placeholders in the generated pdf.

There's a safety check, though, so if the 'final' option is
set in the documentclass, LaTeX throws an error if files are
missing."